Add files via upload
Fixed error where the game state didn't switch to the space shooter after completing flashcards or importing a flashcards.txt file. The assets would not load, as I wrote: ImageIO.read(new File(filePath)). You essentially are asking Java to look for a real file on the local file system, within a directory. However, that only works in an IDE, because all those files exist on the disk. When you package a Java file into. jar, there are no longer directories, but one JAR archive. So, the best practice is to use: ImageIO.read(getClass().getResourceAsStream(filePath)). This is because it tells Java to look inside the JAR's internal structure and find the file in the class path.